elm_gen{,grid,list}: add support for the reusable_content_get callback.
This is a fairly new callback for genlists and gengrids. As far as I can
tell it makes it possible to avoid the creation of new item objects by
recycling an item that has recently gone of-screen.

I've found it very useful for prepending items to a genlist and I haven't
found drawbacks to it yet but I also haven't been able to understand why
it needed the flexibility of a callback instead of being always enabled
and my callbacks have always been very trivial:

    func_reusable_content_get = (fun _obj _part _old -> Some _old);

The callback receives an "old" object and returns an Evas.obj option. If
the return value is Some o, the object "o" will be used instead of creating
a new item object; if it is None, a new item object will be created.